Restore persisted state after data is fully loaded (#463)
closes rdar://r101539534

* fix: persist state after loading
* refactor: create BaseNavigatorCard and LoadingNavigatorCard
* fix: set common head to BaseNavigatorCard
