+++ src/components/Navigator/NavigatorCard.vue
@@ -9,119 +9,92 @@
 -->
 
 <template>
-  <div class="navigator-card" :class="{ 'filter-on-top': renderFilterOnTop }">
-    <div class="navigator-card-full-height">
-      <div class="navigator-card-inner">
-        <div class="head-wrapper">
-          <div class="head-inner">
-            <button
-              aria-label="Close documentation navigator"
-              :id="SIDEBAR_HIDE_BUTTON_ID"
-              class="close-card"
-              :class="{ 'hide-on-large': !allowHiding }"
-              @click="handleHideClick"
-            >
-              <SidenavIcon class="icon-inline close-icon" />
-            </button>
-            <Reference
-              class="navigator-head"
-              :url="technologyPath"
-              :id="INDEX_ROOT_KEY"
-              @click.alt.native.prevent="toggleAllNodes"
-            >
-              <h2 class="card-link">
-                {{ technology }}
-              </h2>
-              <Badge v-if="isTechnologyBeta" variant="beta" />
-            </Reference>
-          </div>
-        </div>
-        <slot name="post-head" />
-        <div
-          class="card-body"
+  <BaseNavigatorCard
+    :class="{ 'filter-on-top': renderFilterOnTop }"
+    v-bind="{
+      technology,
+      isTechnologyBeta,
+      technologyPath,
+    }"
+    @close="$emit('close')"
+    @head-click-alt="toggleAllNodes"
+  >
+    <template #body="{ className }">
+      <slot name="post-head" />
+      <div
+        :class="className"
+        @keydown.alt.up.capture.prevent="focusFirst"
+        @keydown.alt.down.capture.prevent="focusLast"
+        @keydown.up.exact.capture.prevent="focusPrev"
+        @keydown.down.exact.capture.prevent="focusNext"
+      >
+        <DynamicScroller
+          v-show="hasNodes"
+          :id="scrollLockID"
+          ref="scroller"
+          class="scroller"
+          aria-label="Documentation Navigator"
+          :items="nodesToRender"
+          :min-item-size="itemSize"
+          emit-update
+          key-field="uid"
+          v-slot="{ item, active, index }"
+          @focusin.native="handleFocusIn"
+          @focusout.native="handleFocusOut"
+          @update="handleScrollerUpdate"
           @keydown.alt.up.capture.prevent="focusFirst"
           @keydown.alt.down.capture.prevent="focusLast"
           @keydown.up.exact.capture.prevent="focusPrev"
           @keydown.down.exact.capture.prevent="focusNext"
         >
-          <template v-if="isLoading">
-            <transition name="delay-visibility" appear>
-              <div class="scroller" aria-hidden="true">
-                <LoadingNavigatorItem
-                  v-for="(row, index) in LOADER_ROWS"
-                  :key="index"
-                  :index="index"
-                  :width="row.width"
-                  :hideNavigatorIcon="row.hideNavigatorIcon"
-                />
-              </div>
-            </transition>
-          </template>
-          <DynamicScroller
-            v-else
-            v-show="hasNodes"
-            :id="scrollLockID"
-            ref="scroller"
-            class="scroller"
-            aria-label="Documentation Navigator"
-            :items="nodesToRender"
-            :min-item-size="itemSize"
-            emit-update
-            key-field="uid"
-            v-slot="{ item, active, index }"
-            @focusin.native="handleFocusIn"
-            @focusout.native="handleFocusOut"
-            @update="handleScrollerUpdate"
-          >
-            <DynamicScrollerItem v-bind="{ active, item, dataIndex: index }">
-              <NavigatorCardItem
-                :item="item"
-                :isRendered="active"
-                :filter-pattern="filterPattern"
-                :is-active="item.uid === activeUID"
-                :is-bold="activePathMap[item.uid]"
-                :expanded="openNodes[item.uid]"
-                :api-change="apiChangesObject[item.path]"
-                :isFocused="focusedIndex === index"
-                :enableFocus="!externalFocusChange"
-                :navigator-references="navigatorReferences"
-                @toggle="toggle"
-                @toggle-full="toggleFullTree"
-                @toggle-siblings="toggleSiblings"
-                @navigate="handleNavigationChange"
-                @focus-parent="focusNodeParent"
-              />
-            </DynamicScrollerItem>
-          </DynamicScroller>
-          <div aria-live="polite" class="visuallyhidden">
-            {{ politeAriaLive }}
-          </div>
-          <div aria-live="assertive" class="no-items-wrapper">
-            <p class="no-items">
-              {{ assertiveAriaLive }}
-            </p>
-          </div>
+          <DynamicScrollerItem v-bind="{ active, item, dataIndex: index }">
+            <NavigatorCardItem
+              :item="item"
+              :isRendered="active"
+              :filter-pattern="filterPattern"
+              :is-active="item.uid === activeUID"
+              :is-bold="activePathMap[item.uid]"
+              :expanded="openNodes[item.uid]"
+              :api-change="apiChangesObject[item.path]"
+              :isFocused="focusedIndex === index"
+              :enableFocus="!externalFocusChange"
+              :navigator-references="navigatorReferences"
+              @toggle="toggle"
+              @toggle-full="toggleFullTree"
+              @toggle-siblings="toggleSiblings"
+              @navigate="handleNavigationChange"
+              @focus-parent="focusNodeParent"
+            />
+          </DynamicScrollerItem>
+        </DynamicScroller>
+        <div aria-live="polite" class="visuallyhidden">
+          {{ politeAriaLive }}
         </div>
-        <div class="filter-wrapper" v-if="!errorFetching">
-          <div class="navigator-filter">
-            <div class="input-wrapper">
-              <FilterInput
-                v-model="filter"
-                :tags="availableTags"
-                :selected-tags.sync="selectedTagsModelValue"
-                placeholder="Filter"
-                :should-keep-open-on-blur="false"
-                :position-reversed="!renderFilterOnTop"
-                :clear-filter-on-tag-select="false"
-                class="filter-component"
-                @clear="clearFilters"
-              />
-            </div>
+        <div aria-live="assertive" class="no-items-wrapper">
+          <p class="no-items">
+            {{ assertiveAriaLive }}
+          </p>
+        </div>
+      </div>
+      <div class="filter-wrapper" v-if="!errorFetching">
+        <div class="navigator-filter">
+          <div class="input-wrapper">
+            <FilterInput
+              v-model="filter"
+              :tags="availableTags"
+              :selected-tags.sync="selectedTagsModelValue"
+              placeholder="Filter"
+              :should-keep-open-on-blur="false"
+              :position-reversed="!renderFilterOnTop"
+              :clear-filter-on-tag-select="false"
+              class="filter-component"
+              @clear="clearFilters"
+            />
           </div>
         </div>
       </div>
-    </div>
-  </div>
+    </template>
+  </BaseNavigatorCard>
 </template>
 
 <script>
@@ -132,21 +105,16 @@ import debounce from 'docc-render/utils/debounce';
 import { sessionStorage } from 'docc-render/utils/storage';
 import {
   INDEX_ROOT_KEY,
-  SIDEBAR_HIDE_BUTTON_ID,
   SIDEBAR_ITEM_SIZE,
 } from 'docc-render/constants/sidebar';
 import { safeHighlightPattern } from 'docc-render/utils/search-utils';
-import LoadingNavigatorItem from 'docc-render/components/Navigator/LoadingNavigatorItem.vue';
 import NavigatorCardItem from 'docc-render/components/Navigator/NavigatorCardItem.vue';
-import SidenavIcon from 'theme/components/Icons/SidenavIcon.vue';
-import Reference from 'docc-render/components/ContentNode/Reference.vue';
+import BaseNavigatorCard from 'docc-render/components/Navigator/BaseNavigatorCard.vue';
 import { TopicTypes } from 'docc-render/constants/TopicTypes';
 import FilterInput from 'docc-render/components/Filter/FilterInput.vue';
 import keyboardNavigation from 'docc-render/mixins/keyboardNavigation';
 import { isEqual, last } from 'docc-render/utils/arrays';
 import { ChangeNames, ChangeNameToType } from 'docc-render/constants/Changes';
-import Badge from 'docc-render/components/Badge.vue';
-import { baseNavOpenSidenavButtonId } from 'docc-render/constants/nav';
 import {
   convertChildrenArrayToObject,
   getAllChildren,
@@ -167,12 +135,6 @@ const FILTER_TAGS = {
   tutorials: 'tutorials',
   articles: 'articles',
 };
-
-const LOADER_ROWS = [
-  { width: '30%', hideNavigatorIcon: true },
-  { width: '80%' },
-  { width: '50%' },
-];
 
 const FILTER_TAGS_TO_LABELS = {
   [FILTER_TAGS.sampleCode]: 'Sample Code',
@@ -219,20 +181,14 @@ export default {
     HIDE_DEPRECATED_TAG,
   },
   components: {
-    Badge,
     FilterInput,
-    SidenavIcon,
-    LoadingNavigatorItem,
     NavigatorCardItem,
     DynamicScroller,
     DynamicScrollerItem,
-    Reference,
+    BaseNavigatorCard,
   },
   props: {
-    technology: {
-      type: String,
-      required: true,
-    },
+    ...BaseNavigatorCard.props,
     children: {
       type: Array,
       required: true,
@@ -244,10 +200,6 @@ export default {
     type: {
       type: String,
       required: true,
-    },
-    technologyPath: {
-      type: String,
-      default: '',
     },
     scrollLockID: {
       type: String,
@@ -262,14 +214,6 @@ export default {
       default: null,
     },
     isTechnologyBeta: {
-      type: Boolean,
-      default: false,
-    },
-    allowHiding: {
-      type: Boolean,
-      default: true,
-    },
-    isLoading: {
       type: Boolean,
       default: false,
     },
@@ -302,13 +246,10 @@ export default {
       NO_CHILDREN,
       ERROR_FETCHING,
       ITEMS_FOUND,
-      LOADER_ROWS,
       allNodesToggled: false,
     };
   },
   computed: {
-    SIDEBAR_HIDE_BUTTON_ID: () => SIDEBAR_HIDE_BUTTON_ID,
-    INDEX_ROOT_KEY: () => INDEX_ROOT_KEY,
     politeAriaLive: ({ hasNodes, nodesToRender }) => {
       if (!hasNodes) return '';
       return [nodesToRender.length, ITEMS_FOUND].join(' ');
@@ -1066,14 +1007,6 @@ export default {
       // we perform an intentional focus change, so no need to set `externalFocusChange` to `true`
       this.focusIndex(parentIndex);
     },
-    async handleHideClick() {
-      this.$emit('close');
-      await this.$nextTick();
-      const trigger = document.getElementById(baseNavOpenSidenavButtonId);
-      if (trigger) {
-        trigger.focus();
-      }
-    },
   },
 };
 </script>
@@ -1082,33 +1015,11 @@ export default {
 @import 'docc-render/styles/_core.scss';
 @import '~vue-virtual-scroller/dist/vue-virtual-scroller.css';
 
-$navigator-card-vertical-spacing: 8px !default;
 // unfortunately we need to hard-code the filter height
 $filter-height: 73px;
 $filter-height-small: 62px;
-$navigator-head-background: var(--color-fill) !default;
-$navigator-head-background-active: var(--color-fill) !default;
-$close-icon-size: 19px;
-$close-icon-padding: 5px;
 
 .navigator-card {
-  --card-vertical-spacing: #{$navigator-card-vertical-spacing};
-  --card-horizontal-spacing: #{$nav-card-horizontal-spacing-large};
-  --nav-filter-horizontal-padding: 30px;
-  --visibility-delay: 1s; // don't show spinner until this much time has passed
-  display: flex;
-  flex-direction: column;
-  min-height: 0;
-  height: calc(var(--app-height) - var(--nav-height, 0px));
-  position: sticky;
-  top: var(--nav-height, 0px);
-
-  @include breakpoint(medium, nav) {
-    height: 100%;
-    position: static;
-    background: var(--color-fill);
-  }
-
   &.filter-on-top {
     .filter-wrapper {
       order: 1;
@@ -1118,66 +1029,6 @@ $close-icon-padding: 5px;
     .card-body {
       order: 2;
     }
-  }
-
-  .navigator-card-full-height {
-    min-height: 0;
-    flex: 1 1 auto;
-  }
-
-  .head-inner {
-    overflow: hidden;
-  }
-
-  .head-wrapper {
-    position: relative;
-    flex: 1 0 auto;
-  }
-
-  .navigator-head {
-    --navigator-head-padding-right: calc(var(--card-horizontal-spacing) * 2 + #{$close-icon-size});
-    padding: 0 var(--navigator-head-padding-right) 0 var(--card-horizontal-spacing);
-    background: $navigator-head-background;
-    border-bottom: 1px solid var(--color-grid);
-    display: flex;
-    align-items: center;
-    height: $nav-height;
-    white-space: nowrap;
-
-    .badge {
-      margin-top: 0;
-    }
-
-    &.router-link-exact-active {
-      background: $navigator-head-background-active;
-
-      .card-link {
-        font-weight: $font-weight-bold;
-      }
-    }
-
-    &:hover {
-      background: var(--color-navigator-item-hover);
-      text-decoration: none;
-    }
-
-    @include safe-area-left-set(padding-left, var(--card-horizontal-spacing));
-    @include safe-area-right-set(padding-right, var(--navigator-head-padding-right));
-
-    @include breakpoint(medium, nav) {
-      justify-content: center;
-      --navigator-head-padding-right: var(--card-horizontal-spacing);
-    }
-
-    @include breakpoint(small, nav) {
-      height: $nav-height-small;
-      padding: 0 $nav-card-horizontal-spacing-large;
-    }
-  }
-
-  .card-icon {
-    width: 19px;
-    height: 19px;
   }
 }
 
@@ -1192,79 +1043,6 @@ $close-icon-padding: 5px;
     min-width: 200px;
     box-sizing: border-box;
   }
-}
-
-.close-card {
-  display: flex;
-  position: absolute;
-  z-index: 1;
-  align-items: center;
-  justify-content: center;
-  right: $nav-padding - rem($close-icon-padding);
-  padding: $close-icon-padding;
-  margin-left: -$close-icon-padding;
-  top: calc(50% - #{$close-icon-size} + #{$close-icon-padding});
-  transition: transform $adjustable-sidebar-hide-transition-duration ease-in 0s, visibility 0s;
-
-  @include breakpoint(medium, nav) {
-    right: unset;
-    top: 0;
-    left: 0;
-    margin: 0;
-    padding: 0 $nav-padding 0 $nav-card-horizontal-spacing-large;
-    height: 100%;
-    @include safe-area-left-set(padding-left, $nav-padding);
-  }
-
-  @include breakpoint(small, nav) {
-    padding-left: $nav-padding-small;
-    padding-right: $nav-padding-small;
-    @include safe-area-left-set(padding-left, $nav-padding-small);
-  }
-
-  .close-icon {
-    width: $close-icon-size;
-    height: $close-icon-size;
-  }
-
-  @include breakpoints-from(large, nav) {
-    &.hide-on-large {
-      display: none;
-    }
-
-    &:hover {
-      border-radius: $nano-border-radius;
-      background: var(--color-fill-gray-quaternary);
-    }
-    // when the navigator is closed on desktop,
-    // move the button so it looks like its going to the nav
-    .sidebar-hidden & {
-      transition: transform $adjustable-sidebar-hide-transition-duration ease-in 0s,
-      visibility 0s linear $adjustable-sidebar-hide-transition-duration;
-      visibility: hidden;
-      // 2x the nav padding, 1px border, and the size of the icon
-      transform: translateX(rem($close-icon-size + 1px) + $nav-padding * 2);
-    }
-  }
-}
-
-.card-body {
-  // right padding is added by the items, so visually the scroller is stuck to the side
-  padding-right: 0;
-  flex: 1 1 auto;
-  min-height: 0;
-  height: 100%;
-  @include breakpoint(medium, nav) {
-    --card-vertical-spacing: 0px;
-  }
-}
-
-.card-link {
-  color: var(--color-text);
-  @include font-styles(body-reduced);
-  font-weight: $font-weight-semibold;
-  overflow: hidden;
-  text-overflow: ellipsis;
 }
 
 .navigator-filter {
@@ -1337,23 +1115,5 @@ $close-icon-padding: 5px;
     flex: 1 0 $filter-height;
     overflow: hidden;
   }
-}
-
-.navigator-card-inner {
-  --nav-card-inner-vertical-offset: 0px;
-  position: sticky;
-  top: var(--nav-height, 0px);
-  height: calc(var(--app-height) - var(--nav-height, 0px) - var(--nav-card-inner-vertical-offset));
-  display: flex;
-  flex-flow: column;
-  @include breakpoint(medium, nav) {
-    position: static;
-    height: 100%;
-  }
-}
-
-.delay-visibility-enter-active {
-  transition: visibility var(--visibility-delay);
-  visibility: hidden;
 }
 </style>
