+++ tests/unit/components/Navigator.spec.js
@@ -145,7 +145,6 @@ describe('Navigator', () => {
       scrollLockID: defaultProps.scrollLockID,
       renderFilterOnTop: defaultProps.renderFilterOnTop,
       errorFetching: false,
-      isLoading: false,
       apiChanges: null,
       allowHiding: true,
       navigatorReferences,
@@ -185,7 +184,6 @@ describe('Navigator', () => {
       scrollLockID: defaultProps.scrollLockID,
       renderFilterOnTop: defaultProps.renderFilterOnTop,
       errorFetching: false,
-      isLoading: false,
       apiChanges: null,
       allowHiding: true,
       navigatorReferences,
